requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
openai
streamlit
beautifulsoup4
requests
python-dotenv
